Sync SPU 0.2.x

+++ build_and_packaging.sh
@@ -1,17 +1,36 @@
 #! /bin/bash
+#
+# Copyright 2022 Ant Group Co., Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 RED='\033[0;31m'
 NC='\033[0m'
 
 function build_spu(){
     echo -e "${RED}Start build wheel package...${NC}"
 
-    docker run --rm --mount type=bind,source="$(pwd)",target=/home/admin/dev/ \
+    if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+        docker run --rm --mount type=bind,source="$(pwd)",target=/home/admin/dev/ \
         -w /home/admin/dev \
         --cap-add=SYS_PTRACE --security-opt seccomp=unconfined \
         --cap-add=NET_ADMIN \
         --privileged=true \
         --entrypoint "./build_wheel_entrypoint.sh" \
-        registry.hub.docker.com/secretflow/spu-gcc11-anolis-dev:latest
+        registry.hub.docker.com/secretflow/sf-gcc11-centos7-amd64-release:latest
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        sh ./build_wheel_entrypoint.sh
+    fi    
 
     (($? != 0)) && { echo -e "${RED}Build exited with non-zero.${NC}"; exit 1; }
